extract GenerateKeyUseCase
